% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bone_list.R
\name{bone_list}
\alias{bone_list}
\title{Sample dataset of bone shapes}
\format{A data frame with 700 rows and 2 columns:
\describe{
  \item{id}{id}
  \item{x}{x- coordinate}
  \item{y}{y- coordinate}
  ...
}}
\description{
A dataset containing 700 x- and y- coordinates. Each group of 100 represents a 2D cross-section of a bone.
}
