% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{data_transform}
\alias{data_transform}
\title{Transforms data into usable format.}
\usage{
data_transform(data)
}
\arguments{
\item{data}{input data}
}
\value{
transformed data
}
\description{
Transforms data into usable format.
}
\examples{

\dontrun{
test_data = bone_list
test_data = t(test_data)
transform_data(test_data)
}

}
\author{
Goutham Swaminathan \email{goutham1220@gmail.com}
}
