% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_exp_coord.R
\name{inverse_exp_coord}
\alias{inverse_exp_coord}
\title{Main internal function for calculating Elastic Distance
Adapted from "fdasrvf" package by  J. Derek Tucker <jdtuck@sandia.gov>}
\usage{
inverse_exp_coord(beta1, beta2, mode = "O", rotated = T)
}
\arguments{
\item{beta1}{first curve}

\item{beta2}{second curve}

\item{mode}{open curve ("O") or closed curve("C")}

\item{rotated}{account for rotation}
}
\description{
Main internal function for calculating Elastic Distance
Adapted from "fdasrvf" package by  J. Derek Tucker <jdtuck@sandia.gov>
}
\author{
Goutham Swaminathan \email{goutham1220@gmail.com}
}
