% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bones_data.R
\name{load_bones_data}
\alias{load_bones_data}
\title{loads test dataframe of bone shapes}
\usage{
load_bones_data()
}
\value{
a sample dataframe of bone shapes (7 bones).
}
\description{
loads test dataframe of bone shapes
}
\examples{

\dontrun{
load_bones_data()
}

}
\author{
Goutham Swaminathan \email{goutham1220@gmail.com}
}
