% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_elastic_distance.R
\name{shape_elastic_distance}
\alias{shape_elastic_distance}
\title{Calculates Elastic distances for all nC2 combinations of a 3-dimensional array of shapes
Adapted from "fdasrvf" package by  J. Derek Tucker <jdtuck@sandia.gov>}
\usage{
shape_elastic_distance(x, mode = "C")
}
\arguments{
\item{x}{3-dimensional array of shapes. Every zth element should be a different shape.}

\item{mode}{Open ("O") or Closed ("C") curves}
}
\value{
A matrix of all calculated distance. Each distance is labeled by the shapes used for calculation.
}
\description{
Calculates Elastic distances for all nC2 combinations of a 3-dimensional array of shapes
Adapted from "fdasrvf" package by  J. Derek Tucker <jdtuck@sandia.gov>
}
\examples{

\dontrun{
bone_shapes = load_bones_data()
bone_shapes = bones_shapes$x
bone_shapes = cbind(bone_shapes, bone_list$y)
bone_1 = bone_shapes[1:100,]
bone_2 = bone_shapes[101:200,]
bone_3 = bone_shapes[201:300,]
bones = array(c(bone_1, bone_2, bone_3), dim = c(100, 2, 3))
shape_elastic_distance(bones, mode = "C")
}

}
\author{
Goutham Swaminathan \email{goutham1220@gmail.com}
}
