% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_proc_distance.R
\name{shape_proc_distance}
\alias{shape_proc_distance}
\title{Adapted from "shapes" package by Ian L. Dryden <ian.dryden@nottingham.ac.uk>
Calculates Procrustes distances for all nC2 combinations of a 3-dimensional array of shapes}
\usage{
shape_proc_distance(x, type = "full", reflect = FALSE)
}
\arguments{
\item{x}{3-dimensional array of shapes. Every zth element should be a different shape.}

\item{type}{string indicating the type of distance; "full" full Procrustes distance,
"partial" partial Procrustes distance,
"Riemannian" Riemannian shape distance,
"sizeandshape" size-and-shape Riemannian/Procrustes distance}

\item{reflect}{Logical. If reflect = TRUE then reflection invariance is included.}
}
\value{
A matrix of all calculated distance. Each distance is labeled by the shapes used for calculation.
}
\description{
Adapted from "shapes" package by Ian L. Dryden <ian.dryden@nottingham.ac.uk>
Calculates Procrustes distances for all nC2 combinations of a 3-dimensional array of shapes
}
\examples{

\dontrun{
bone_shapes = load_bones_data()
bone_shapes = bones_shapes$x
bone_shapes = cbind(bone_shapes, bone_list$y)
bone_1 = bone_shapes[1:100,]
bone_2 = bone_shapes[101:200,]
bone_3 = bone_shapes[201:300,]
bones = array(c(bone_1, bone_2, bone_3), dim = c(100, 2, 3))
shape_proc_distance(bones, type = "full")
}

}
\author{
Goutham Swaminathan \email{goutham1220@gmail.com}
}
