% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R
\name{data_split}
\alias{data_split}
\title{split a matrix into desired 3-D array at set interval.}
\usage{
data_split(matrix, interval)
}
\arguments{
\item{matrix}{input matrix}

\item{interval}{interval for splitting of matrix}
}
\value{
3-D array
}
\description{
split a matrix into desired 3-D array at set interval.
}
\examples{

\dontrun{
test_data = bone_list
bones = data_split(test_data, 100)
}

}
\author{
Goutham Swaminathan \email{goutham1220@gmail.com}
}
